+++ crates/utils/transaction-generator/tests/message_generators/main.rs
@@ -216,7 +216,13 @@ fn setup_harness<
             ValueSetterDiscriminants::SetValue,
             ValueSetterDiscriminants::SetManyValues,
         ]),
-        max_value_setter_vec_len,
+        sov_transaction_generator::generators::value_setter::ValueSetterGeneratorOptions {
+            maximum_vec_length: max_value_setter_vec_len,
+            min_and_max_number_of_individual_state_operations: (0, 0),
+            min_and_max_number_of_new_values_for_heavy_state: (0, 0),
+            min_and_max_number_of_iterations_for_cpu_heavy_operation: (0, 0),
+            max_heavy_state_size: 0,
+        },
         admin.private_key.clone(),
     ));