+++ examples/demo-rollup/autogenerated.rs
@@ -1,4 +1,4 @@
-pub const CHAIN_HASH: [u8; 32] = [180, 242, 252, 125, 36, 177, 131, 70, 141, 167, 100, 247, 164, 146, 220, 231, 82, 72, 74, 202, 14, 27, 44, 69, 119, 253, 231, 9, 170, 188, 170, 222];
+pub const CHAIN_HASH: [u8; 32] = [146, 161, 27, 145, 7, 28, 75, 233, 37, 207, 169, 212, 59, 243, 252, 4, 105, 8, 48, 123, 232, 101, 139, 36, 90, 63, 90, 70, 135, 132, 217, 199];
 
 #[allow(dead_code)]
 pub const SCHEMA_JSON: &str = r#"{
@@ -98,7 +98,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "details",
             "silent": false,
             "value": {
-              "ByIndex": 109
+              "ByIndex": 112
             },
             "doc": ""
           }
@@ -182,7 +182,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 3,
             "template": null,
             "value": {
-              "ByIndex": 40
+              "ByIndex": 43
             }
           },
           {
@@ -190,7 +190,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 4,
             "template": null,
             "value": {
-              "ByIndex": 42
+              "ByIndex": 45
             }
           },
           {
@@ -198,7 +198,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 5,
             "template": null,
             "value": {
-              "ByIndex": 47
+              "ByIndex": 50
             }
           },
           {
@@ -206,7 +206,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 6,
             "template": null,
             "value": {
-              "ByIndex": 51
+              "ByIndex": 54
             }
           },
           {
@@ -214,7 +214,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 7,
             "template": null,
             "value": {
-              "ByIndex": 56
+              "ByIndex": 59
             }
           },
           {
@@ -222,7 +222,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 8,
             "template": null,
             "value": {
-              "ByIndex": 57
+              "ByIndex": 60
             }
           },
           {
@@ -230,7 +230,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 9,
             "template": null,
             "value": {
-              "ByIndex": 58
+              "ByIndex": 61
             }
           },
           {
@@ -238,7 +238,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 10,
             "template": null,
             "value": {
-              "ByIndex": 59
+              "ByIndex": 62
             }
           },
           {
@@ -246,7 +246,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 11,
             "template": null,
             "value": {
-              "ByIndex": 85
+              "ByIndex": 88
             }
           },
           {
@@ -254,7 +254,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 12,
             "template": null,
             "value": {
-              "ByIndex": 88
+              "ByIndex": 91
             }
           }
         ],
@@ -873,8 +873,32 @@ pub const SCHEMA_JSON: &str = r#"{
             }
           },
           {
-            "name": "Panic",
+            "name": "ReadAndSetManyIndividualValues",
             "discriminant": 3,
+            "template": null,
+            "value": {
+              "ByIndex": 40
+            }
+          },
+          {
+            "name": "ReadAndSetHeavyState",
+            "discriminant": 4,
+            "template": null,
+            "value": {
+              "ByIndex": 41
+            }
+          },
+          {
+            "name": "RunCPUHeavyOperation",
+            "discriminant": 5,
+            "template": null,
+            "value": {
+              "ByIndex": 42
+            }
+          },
+          {
+            "name": "Panic",
+            "discriminant": 6,
             "template": null,
             "value": null
           }
@@ -989,13 +1013,118 @@ pub const SCHEMA_JSON: &str = r#"{
       }
     },
     {
+      "Struct": {
+        "type_name": "__SovVirtualWallet_CallMessage_ReadAndSetManyIndividualValues",
+        "template": null,
+        "peekable": false,
+        "fields": [
+          {
+            "display_name": "number_of_operations",
+            "silent": false,
+            "value": {
+              "Immediate": {
+                "Integer": [
+                  "u64",
+                  "Decimal"
+                ]
+              }
+            },
+            "doc": ""
+          },
+          {
+            "display_name": "salt",
+            "silent": false,
+            "value": {
+              "Immediate": {
+                "Integer": [
+                  "u64",
+                  "Decimal"
+                ]
+              }
+            },
+            "doc": ""
+          }
+        ]
+      }
+    },
+    {
+      "Struct": {
+        "type_name": "__SovVirtualWallet_CallMessage_ReadAndSetHeavyState",
+        "template": null,
+        "peekable": false,
+        "fields": [
+          {
+            "display_name": "number_of_new_values",
+            "silent": false,
+            "value": {
+              "Immediate": {
+                "Integer": [
+                  "u64",
+                  "Decimal"
+                ]
+              }
+            },
+            "doc": ""
+          },
+          {
+            "display_name": "max_heavy_state_size",
+            "silent": false,
+            "value": {
+              "Immediate": {
+                "Integer": [
+                  "u64",
+                  "Decimal"
+                ]
+              }
+            },
+            "doc": ""
+          },
+          {
+            "display_name": "salt",
+            "silent": false,
+            "value": {
+              "Immediate": {
+                "Integer": [
+                  "u64",
+                  "Decimal"
+                ]
+              }
+            },
+            "doc": ""
+          }
+        ]
+      }
+    },
+    {
+      "Struct": {
+        "type_name": "__SovVirtualWallet_CallMessage_RunCPUHeavyOperation",
+        "template": null,
+        "peekable": false,
+        "fields": [
+          {
+            "display_name": "iterations",
+            "silent": false,
+            "value": {
+              "Immediate": {
+                "Integer": [
+                  "u64",
+                  "Decimal"
+                ]
+              }
+            },
+            "doc": ""
+          }
+        ]
+      }
+    },
+    {
       "Tuple": {
         "template": null,
         "peekable": false,
         "fields": [
           {
             "value": {
-              "ByIndex": 41
+              "ByIndex": 44
             },
             "silent": false,
             "doc": ""
@@ -1017,7 +1146,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 43
+              "ByIndex": 46
             },
             "silent": false,
             "doc": ""
@@ -1034,7 +1163,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 0,
             "template": null,
             "value": {
-              "ByIndex": 44
+              "ByIndex": 47
             }
           },
           {
@@ -1054,7 +1183,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 3,
             "template": null,
             "value": {
-              "ByIndex": 45
+              "ByIndex": 48
             }
           },
           {
@@ -1068,7 +1197,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 5,
             "template": null,
             "value": {
-              "ByIndex": 46
+              "ByIndex": 49
             }
           }
         ],
@@ -1127,7 +1256,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 48
+              "ByIndex": 51
             },
             "silent": false,
             "doc": ""
@@ -1144,7 +1273,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 0,
             "template": null,
             "value": {
-              "ByIndex": 49
+              "ByIndex": 52
             }
           },
           {
@@ -1152,7 +1281,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 1,
             "template": null,
             "value": {
-              "ByIndex": 50
+              "ByIndex": 53
             }
           },
           {
@@ -1202,7 +1331,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 52
+              "ByIndex": 55
             },
             "silent": false,
             "doc": ""
@@ -1219,7 +1348,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 0,
             "template": null,
             "value": {
-              "ByIndex": 53
+              "ByIndex": 56
             }
           }
         ],
@@ -1233,7 +1362,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 54
+              "ByIndex": 57
             },
             "silent": false,
             "doc": ""
@@ -1248,7 +1377,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 55
+              "ByIndex": 58
             },
             "silent": false,
             "doc": ""
@@ -1283,7 +1412,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 41
+              "ByIndex": 44
             },
             "silent": false,
             "doc": ""
@@ -1298,7 +1427,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 41
+              "ByIndex": 44
             },
             "silent": false,
             "doc": ""
@@ -1313,7 +1442,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 41
+              "ByIndex": 44
             },
             "silent": false,
             "doc": ""
@@ -1328,7 +1457,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 60
+              "ByIndex": 63
             },
             "silent": false,
             "doc": ""
@@ -1345,7 +1474,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 0,
             "template": null,
             "value": {
-              "ByIndex": 61
+              "ByIndex": 64
             }
           },
           {
@@ -1353,7 +1482,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 1,
             "template": null,
             "value": {
-              "ByIndex": 74
+              "ByIndex": 77
             }
           },
           {
@@ -1361,7 +1490,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 2,
             "template": null,
             "value": {
-              "ByIndex": 75
+              "ByIndex": 78
             }
           }
         ],
@@ -1378,7 +1507,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "policy",
             "silent": false,
             "value": {
-              "ByIndex": 62
+              "ByIndex": 65
             },
             "doc": ""
           }
@@ -1395,7 +1524,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "default_payee_policy",
             "silent": false,
             "value": {
-              "ByIndex": 63
+              "ByIndex": 66
             },
             "doc": ""
           },
@@ -1403,7 +1532,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "payees",
             "silent": false,
             "value": {
-              "ByIndex": 69
+              "ByIndex": 72
             },
             "doc": ""
           },
@@ -1419,7 +1548,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "authorized_sequencers",
             "silent": false,
             "value": {
-              "ByIndex": 71
+              "ByIndex": 74
             },
             "doc": ""
           }
@@ -1435,7 +1564,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 0,
             "template": null,
             "value": {
-              "ByIndex": 64
+              "ByIndex": 67
             }
           },
           {
@@ -1474,7 +1603,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "max_gas_price",
             "silent": false,
             "value": {
-              "ByIndex": 65
+              "ByIndex": 68
             },
             "doc": ""
           },
@@ -1482,7 +1611,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "transaction_limit",
             "silent": false,
             "value": {
-              "ByIndex": 68
+              "ByIndex": 71
             },
             "doc": ""
           }
@@ -1492,7 +1621,7 @@ pub const SCHEMA_JSON: &str = r#"{
     {
       "Option": {
         "value": {
-          "ByIndex": 66
+          "ByIndex": 69
         }
       }
     },
@@ -1506,7 +1635,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "value",
             "silent": false,
             "value": {
-              "ByIndex": 67
+              "ByIndex": 70
             },
             "doc": ""
           }
@@ -1536,7 +1665,7 @@ pub const SCHEMA_JSON: &str = r#"{
     {
       "Vec": {
         "value": {
-          "ByIndex": 70
+          "ByIndex": 73
         }
       }
     },
@@ -1554,7 +1683,7 @@ pub const SCHEMA_JSON: &str = r#"{
           },
           {
             "value": {
-              "ByIndex": 63
+              "ByIndex": 66
             },
             "silent": false,
             "doc": ""
@@ -1577,7 +1706,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 1,
             "template": null,
             "value": {
-              "ByIndex": 72
+              "ByIndex": 75
             }
           }
         ],
@@ -1591,7 +1720,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 73
+              "ByIndex": 76
             },
             "silent": false,
             "doc": ""
@@ -1641,7 +1770,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "update",
             "silent": false,
             "value": {
-              "ByIndex": 76
+              "ByIndex": 79
             },
             "doc": ""
           }
@@ -1658,7 +1787,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "sequencer_update",
             "silent": false,
             "value": {
-              "ByIndex": 77
+              "ByIndex": 80
             },
             "doc": ""
           },
@@ -1666,7 +1795,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "updaters_to_add",
             "silent": false,
             "value": {
-              "ByIndex": 82
+              "ByIndex": 85
             },
             "doc": ""
           },
@@ -1674,7 +1803,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "updaters_to_remove",
             "silent": false,
             "value": {
-              "ByIndex": 82
+              "ByIndex": 85
             },
             "doc": ""
           },
@@ -1682,7 +1811,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "payee_policies_to_set",
             "silent": false,
             "value": {
-              "ByIndex": 83
+              "ByIndex": 86
             },
             "doc": ""
           },
@@ -1690,7 +1819,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "payee_policies_to_delete",
             "silent": false,
             "value": {
-              "ByIndex": 82
+              "ByIndex": 85
             },
             "doc": ""
           },
@@ -1698,7 +1827,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "default_policy",
             "silent": false,
             "value": {
-              "ByIndex": 84
+              "ByIndex": 87
             },
             "doc": ""
           }
@@ -1708,7 +1837,7 @@ pub const SCHEMA_JSON: &str = r#"{
     {
       "Option": {
         "value": {
-          "ByIndex": 78
+          "ByIndex": 81
         }
       }
     },
@@ -1727,7 +1856,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 1,
             "template": null,
             "value": {
-              "ByIndex": 79
+              "ByIndex": 82
             }
           }
         ],
@@ -1741,7 +1870,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 80
+              "ByIndex": 83
             },
             "silent": false,
             "doc": ""
@@ -1759,7 +1888,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "to_add",
             "silent": false,
             "value": {
-              "ByIndex": 81
+              "ByIndex": 84
             },
             "doc": ""
           },
@@ -1767,7 +1896,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "to_remove",
             "silent": false,
             "value": {
-              "ByIndex": 81
+              "ByIndex": 84
             },
             "doc": ""
           }
@@ -1777,7 +1906,7 @@ pub const SCHEMA_JSON: &str = r#"{
     {
       "Option": {
         "value": {
-          "ByIndex": 73
+          "ByIndex": 76
         }
       }
     },
@@ -1791,14 +1920,14 @@ pub const SCHEMA_JSON: &str = r#"{
     {
       "Option": {
         "value": {
-          "ByIndex": 69
+          "ByIndex": 72
         }
       }
     },
     {
       "Option": {
         "value": {
-          "ByIndex": 63
+          "ByIndex": 66
         }
       }
     },
@@ -1809,7 +1938,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 86
+              "ByIndex": 89
             },
             "silent": false,
             "doc": ""
@@ -1827,7 +1956,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "rlp",
             "silent": false,
             "value": {
-              "ByIndex": 87
+              "ByIndex": 90
             },
             "doc": ""
           }
@@ -1862,7 +1991,7 @@ pub const SCHEMA_JSON: &str = r#"{
         "fields": [
           {
             "value": {
-              "ByIndex": 89
+              "ByIndex": 92
             },
             "silent": false,
             "doc": ""
@@ -1879,7 +2008,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 0,
             "template": null,
             "value": {
-              "ByIndex": 90
+              "ByIndex": 93
             }
           },
           {
@@ -1887,7 +2016,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 1,
             "template": null,
             "value": {
-              "ByIndex": 91
+              "ByIndex": 94
             }
           },
           {
@@ -1895,7 +2024,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 2,
             "template": null,
             "value": {
-              "ByIndex": 93
+              "ByIndex": 96
             }
           },
           {
@@ -1903,7 +2032,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 3,
             "template": null,
             "value": {
-              "ByIndex": 94
+              "ByIndex": 97
             }
           },
           {
@@ -1911,7 +2040,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 4,
             "template": null,
             "value": {
-              "ByIndex": 95
+              "ByIndex": 98
             }
           },
           {
@@ -1919,7 +2048,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 5,
             "template": null,
             "value": {
-              "ByIndex": 96
+              "ByIndex": 99
             }
           },
           {
@@ -1933,7 +2062,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 7,
             "template": null,
             "value": {
-              "ByIndex": 97
+              "ByIndex": 100
             }
           },
           {
@@ -1941,7 +2070,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 8,
             "template": null,
             "value": {
-              "ByIndex": 98
+              "ByIndex": 101
             }
           },
           {
@@ -1955,7 +2084,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 10,
             "template": null,
             "value": {
-              "ByIndex": 99
+              "ByIndex": 102
             }
           },
           {
@@ -1963,7 +2092,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 11,
             "template": null,
             "value": {
-              "ByIndex": 100
+              "ByIndex": 103
             }
           },
           {
@@ -1971,7 +2100,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 12,
             "template": null,
             "value": {
-              "ByIndex": 101
+              "ByIndex": 104
             }
           },
           {
@@ -1979,7 +2108,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 13,
             "template": null,
             "value": {
-              "ByIndex": 108
+              "ByIndex": 111
             }
           }
         ],
@@ -2057,7 +2186,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "content",
             "silent": false,
             "value": {
-              "ByIndex": 92
+              "ByIndex": 95
             },
             "doc": ""
           }
@@ -2315,7 +2444,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "pre",
             "silent": false,
             "value": {
-              "ByIndex": 102
+              "ByIndex": 105
             },
             "doc": ""
           },
@@ -2323,7 +2452,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "post",
             "silent": false,
             "value": {
-              "ByIndex": 102
+              "ByIndex": 105
             },
             "doc": ""
           }
@@ -2333,14 +2462,14 @@ pub const SCHEMA_JSON: &str = r#"{
     {
       "Option": {
         "value": {
-          "ByIndex": 103
+          "ByIndex": 106
         }
       }
     },
     {
       "Vec": {
         "value": {
-          "ByIndex": 104
+          "ByIndex": 107
         }
       }
     },
@@ -2353,7 +2482,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 0,
             "template": null,
             "value": {
-              "ByIndex": 105
+              "ByIndex": 108
             }
           },
           {
@@ -2361,7 +2490,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 1,
             "template": null,
             "value": {
-              "ByIndex": 106
+              "ByIndex": 109
             }
           },
           {
@@ -2369,7 +2498,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "discriminant": 2,
             "template": null,
             "value": {
-              "ByIndex": 107
+              "ByIndex": 110
             }
           }
         ],
@@ -2521,7 +2650,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "max_priority_fee_bips",
             "silent": false,
             "value": {
-              "ByIndex": 110
+              "ByIndex": 113
             },
             "doc": ""
           },
@@ -2608,7 +2737,7 @@ pub const SCHEMA_JSON: &str = r#"{
             "display_name": "details",
             "silent": false,
             "value": {
-              "ByIndex": 109
+              "ByIndex": 112
             },
             "doc": ""
           }
@@ -2618,7 +2747,7 @@ pub const SCHEMA_JSON: &str = r#"{
   ],
   "root_type_indices": [
     0,
-    111,
+    114,
     6,
     14
   ],
@@ -3004,6 +3133,15 @@ pub const SCHEMA_JSON: &str = r#"{
           "name": "assert_visible_slot_number"
         },
         {
+          "name": "read_and_set_many_individual_values"
+        },
+        {
+          "name": "read_and_set_heavy_state"
+        },
+        {
+          "name": "run_cpu_heavy_operation"
+        },
+        {
           "name": "panic"
         }
       ]
@@ -3040,6 +3178,39 @@ pub const SCHEMA_JSON: &str = r#"{
       "fields_or_variants": [
         {
           "name": "expected_visible_slot_number"
+        }
+      ]
+    },
+    {
+      "name": "__SovVirtualWallet_CallMessage_ReadAndSetManyIndividualValues",
+      "fields_or_variants": [
+        {
+          "name": "number_of_operations"
+        },
+        {
+          "name": "salt"
+        }
+      ]
+    },
+    {
+      "name": "__SovVirtualWallet_CallMessage_ReadAndSetHeavyState",
+      "fields_or_variants": [
+        {
+          "name": "number_of_new_values"
+        },
+        {
+          "name": "max_heavy_state_size"
+        },
+        {
+          "name": "salt"
+        }
+      ]
+    },
+    {
+      "name": "__SovVirtualWallet_CallMessage_RunCPUHeavyOperation",
+      "fields_or_variants": [
+        {
+          "name": "iterations"
         }
       ]
     },